ENH: completely re-write the implementation
Closes #2 (by adopting a vendored version of the superqt implementation)

Closes #1 (because the teleporter no longer exists)

In the (old) teleporter version we would create one QObject per process which
had several (very custom) signals on it for each of the things we need to do
that needs to be created on the correct thread.  In the (new) superqt method we
cerate a new QObject per call and then ship it to the correct thread at call
time.